Create Replay Service
Create an Enum for Replay State
Create a Replay Service which stores recorded commands for replay.
Create Replay Service in GameService.
Create a method to update this stack.
Create a method to execute the next command in replay stack if the replay state is active.
Add a delay of 1 second before executing the next command.